feat: Rename project from pay_evm to kanaripay and update related configurations
