+++ android/app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
     <uses-permission android:name="android.permission.USE_FINGERPRINT" />
     
     <application
-        android:label="pay_evm"
+        android:label="KanariPay"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
