+++ lib/screenpage/CreateWalletScreen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:pay_evm/utils/custom_widgets.dart';
-import 'package:pay_evm/screenpage/AppLockScreen.dart';
+import 'package:kanaripay/utils/custom_widgets.dart';
+import 'package:kanaripay/screenpage/AppLockScreen.dart';
 import '../services/wallet_service.dart';
 import '../utils/app_theme.dart';
 import '../main.dart';
